CRUD for ADMIN, ADMIN UI, Enabled delete for Seller

+++ GetItEasy/src/main/java/com/zensar/repository/UserDetailsRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.query.Param;
 
 import com.zensar.beans.UserDetails;
 
-public interface UserDetailsRepository extends JpaRepository<UserDetails, Long> {
+public interface UserDetailsRepository extends JpaRepository<UserDetails, Integer> {
